requirements.txt
Django==4.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.6
coreapi==2.3.3
djangorestframework-simplejwt==5.2.2
django-mptt==0.14.0
django-filter==23.3
coverage==7.3.1
flake8==6.1.0
gunicorn==21.2.0
